practice/requirements.txt
aiohttp==3.12.14; python_version >= '3.9'
aiosignal==1.4.0; python_version >= '3.9'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
attrs==25.3.0; python_version >= '3.8'
blinker==1.9.0; python_version >= '3.9'
certifi==2025.1.31; python_version >= '3.8'
charset-normalizer==3.4.0; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
colorama==0.4.6; python_version >= '3.7'
dataclasses-json==0.6.7; python_version >= '3.8'
faiss-cpu==1.9.0; python_version >= '3.8'
fastapi==0.115.4; python_version >= '3.8'
frozenlist==1.5.0; python_version >= '3.9'
groq==0.11.0; python_version >= '3.8'
httpx==0.27.2; python_version >= '3.8'
idna==3.10; python_version >= '3.8'
jinja2==3.1.4; python_version >= '3.7'
joblib==1.4.2; python_version >= '3.8'
langchain==0.3.7; python_version >= '3.8'
langchain-community==0.3.7; python_version >= '3.8'
langchain-core==0.3.17; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.7'
numpy==1.26.4; python_version >= '3.9'
openai==1.52.1; python_version >= '3.8'
packaging==24.1; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.8'
pydantic==2.9.2; python_version >= '3.8'
pydantic-core==2.23.4; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
scikit-learn==1.5.2; python_version >= '3.8'
scipy==1.14.1; python_version >= '3.8'
sentence-transformers==3.2.1; python_version >= '3.8'
starlette==0.40.0; python_version >= '3.8'
streamlit==1.39.0; python_version >= '3.8'
tenacity==9.0.0; python_version >= '3.8'
tqdm==4.66.5; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
urllib3==2.2.3; python_version >= '3.8'
uvicorn==0.32.0; python_version >= '3.8'
yarl==1.18.3; python_version >= '3.9'
